Rename git-change to git-gerrit
The name git-change is already taken:

    https://github.com/Nextdoor/git-change

I didn' want to change it because I think 'git-change' is much more
meaningful given that most of its subcommands can be thought of as
'methods' of a 'change' object.

The name 'git-gerrit' was my first choice and I'm going back to it
since, so far, I couldn't find it used anyware.

Signed-off-by: Gustavo L. de M. Chaves <gnustavo@cpan.org>