Simplify configuration
It was too much work trying to grok Gerrit's configuration from a git
remote. Instead, git-gerrit needs three configuration variables to be
set: 'baseurl', 'username', and 'project'. If not all of them are set
git-gerrit dies with a message instructing the user how to set
them. The message was inspired in the message git shows when the
variables core.user and core.email aren't set.

Signed-off-by: Gustavo L. de M. Chaves <gnustavo@cpan.org>