Accept multi-valued variables in the configuration
Inspired by the Git::Hooks module implementation, I made the %Config
hash map variable names to array-refs of values. This way it is
possible to have more than one value associated with a variable and
use the list of all values when it makes sense, or just the last one.

Signed-off-by: Gustavo L. de M. Chaves <gnustavo@cpan.org>