Require git-gerrit.remote variable and make git-gerrit.username optional
I can't avoid knowing the git remote referring to the Gerrit
project. If I use the URL instead of the remote name in 'git pull' the
remote branches aren't updated as they should.

The username shouldn't be required because if it's not known it will
be asked for by 'git credential'.

Signed-off-by: Gustavo L. de M. Chaves <gnustavo@cpan.org>