Remove the '--editor' option from sub-commands
The --editor option does not exist anymore. The Git editor will always
be invoked if the --message isn't present. This is similar to
git-commit's interface. It's still possible to not add any message by
specifying an empty message (e.g., '--message ""') to the
sub-commands. This means that git-gerrit doesn't die anymore if empty
messages are specified.

Signed-off-by: Gustavo L. de M. Chaves <gnustavo@cpan.org>