Turn git-gerrit into a proper CPAN distribution
I followed the suggestion at
http://elliotlovesperl.com/2009/11/23/how-to-structure-perl-programs/
and turned the git-gerrit script into a module called Git::Gerrit with
a 'run' exported routine. Now, the script just uses the module and
invokes 'run'. This allowed me to turn it into a proper CPAN
distribution, making it easier to install and updated it.

Signed-off-by: Gustavo L. de M. Chaves <gnustavo@cpan.org>