Avoid open '|-' construct which isn't supported on Windows
The credential_description routine that returned a string was renamed
to credential_description_file and now returns a tuple with a
File::Temp object and the filename of the temporary file which
contains the credential description. This way we can invoke the 'git
credential' command feeding it from the temporary file instead of
printing to its stdin, which isn't supported on Windows.

Signed-off-by: Gustavo L. de M. Chaves <gnustavo@cpan.org>