Fix vote output in 'show'
Contrary to what I thought, the 'permited_labels' component of a
change does not contain the list of all allowed labels. It comes empty
to the changes I saw. So, I had to traverse the 'labels' component to
grok all label and reviewer names once and then traverse it a second
time to output the values.

The formatting is horrible for now. I'll work on it later.

Signed-off-by: Gustavo L. de M. Chaves <gnustavo@cpan.org>