Use alternative methods to grok credentials for pre-1.8 Gits
The git-credential command was implemented on Git 1.8. Git-gerrit now
tries alternative methods to grok credentials if it can't grok then
via git-credential, in this order:

* From the userinfo part of git-gerrit.baseurl.
* From a .netrc file.
* Prompting the user.

The last two options require non-standard Perl modules: Net::Netrc and
Term::Prompt. If they aren't available, they're simply not used. It's
not an error.