Split the config routine into two: config and configs
The config routine returned a scalar or a list depending on its
calling context. But in most places it was called in a natural list
context when we wanted a scalar, forcing us to call it inside a
scalar().

Now we have two routines: config and configs. The first always returns
a scalar while the second always returns a list.