Remove 'push' options from the documentation headings
I substituted the generic "OPTIONS*" for the explicit list of options
from the section heading of the push sub-command because Pod::HTML
wasn't producing the index entry for 'git gerrit push', since it was
too large.

Now the complete list of options for push can be seen at a glance in
the SYNOPSIS and also in their documention in each sub-command
section.