Implement the global option --noop
This option is for "no operation". With it git-gerrit do not execute any
command that will change anything. This means, any Git command affecting
your working tree or any Gerrit REST call affecting the change status on
the server. Use it along with --debug to see what git-gerrit would do.